# Adds perf test for xml serialiser (and memento-gem)