The output on my machine is shown below. This is with a deliberately big entity (100 config keys and 100 sensor values).

```
2017-03-20 14:38:06,468 INFO  test=PerformanceTestDescriptor{summary=mementoSerializer.serializeEntityMemento(size=65850chars), iterations=1000, warmupIterations=100, job=org.apache.brooklyn.core.mgmt.persist.XmlMementoSerializerPerformanceTest$1@28e5dd6d, minAcceptablePerSecond=1.0}; result=PerformanceTestResult{duration=2s 350ms, warmup=574ms, iterations=1000, warmupIterations=100, ratePerSecond=425.531914893617, histogram={1.049ms-2.097ms: 119, 2.097ms-4.194ms: 876, 4.194ms-8.389ms: 5}, cpuTotalFraction=1.0238435008590299, cpuSampleFractions=[1.038314, 1.016156843156843], minAcceptablePerSecond=1.0}
```
and:
```
2017-03-20 14:37:21,040 INFO  test=PerformanceTestDescriptor{summary=mementosGenerators.newBasicMemento(entity), iterations=100, warmupIterations=10, job=org.apache.brooklyn.core.mgmt.rebind.dto.MementoGeneratorsPerformanceTest$1@12fc1b3, minAcceptablePerSecond=1.0}; result=PerformanceTestResult{duration=85ms, warmup=45ms, iterations=100, warmupIterations=10, ratePerSecond=1176.4705882352941, histogram={524.3us-1.049ms: 84, 1.049ms-2.097ms: 16}, cpuTotalFraction=2.5921956613806407, cpuSampleFractions=[], minAcceptablePerSecond=1.0}
```