@tbouron 2017-03-28
What is the use case for this @aledsage ?