@aledsage 2017-03-28
@tbouron I changed the xml-serializer configuration (see https://github.com/apache/brooklyn-server/pull/593#discussion_r105950198) and didn't know what the performance implications would be. I wrote this test so that we have something to measure the actual performance of serialization.

Long term, I'd like us to write the results of these low-level performance tests somewhere (so we can compare the results across different releases etc). However, we've not dedicated any effort to that yet. When we do, performance tests like this will become more useful.