@tbouron 2017-03-28
@aledsage Very cool, didn't know we had that in Brooklyn! Good idea to generalise this, would be very helpful.

Tested locally, got similar results as yours:
```
2017-03-28 15:05:36,826 INFO  test=PerformanceTestDescriptor{summary=mementoSerializer.serializeEntityMemento(size=65850chars), iterations=1000, warmupIterations=100, job=org.apache.brooklyn.core.mgmt.persist.XmlMementoSerializerPerformanceTest$1@6e5bfdfc, minAcceptablePerSecond=1.0}; result=PerformanceTestResult{duration=2s 273ms, warmup=726ms, iterations=1000, warmupIterations=100, ratePerSecond=439.94720633523974, histogram={1.049ms-2.097ms: 496, 2.097ms-4.194ms: 499, 4.194ms-8.389ms: 5}, cpuTotalFraction=1.2468492509767293, cpuSampleFractions=[1.540913260219342, 1.0191804586241275], minAcceptablePerSecond=1.0}
```
and
```
2017-03-28 15:08:27,137 INFO  test=PerformanceTestDescriptor{summary=mementosGenerators.newBasicMemento(entity), iterations=100, warmupIterations=10, job=org.apache.brooklyn.core.mgmt.rebind.dto.MementoGeneratorsPerformanceTest$1@5b11a194, minAcceptablePerSecond=1.0}; result=PerformanceTestResult{duration=40ms, warmup=26ms, iterations=100, warmupIterations=10, ratePerSecond=2500.0, histogram={262.1us-524.3us: 92, 524.3us-1.049ms: 7, 1.049ms-2.097ms: 1}, cpuTotalFraction=4.137444089100175, cpuSampleFractions=[], minAcceptablePerSecond=1.0}
```

LGTM